backend/pyproject.toml
[tool.poetry]
name = "sop-creator-backend"
version = "1.0.0"
description = "SOP Creator Backend API"
authors = ["SOP Creator Team"]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.9"
websockets = "^12.0"

# Database
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"

# Task Queue
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.0"

# AI Integration
anthropic = "^0.31.0"
faster-whisper = "^1.0.0"  # Local Whisper model for speech-to-text (no API key needed, CPU optimized)

# Video Processing
scenedetect = {extras = ["opencv"], version = "^0.6.4"}
ffmpeg-python = "^0.2.0"
opencv-python-headless = "^4.10.0"
Pillow = "^10.4.0"

# External Integrations
google-api-python-client = "^2.137.0"
google-auth-oauthlib = "^1.2.0"
notion-client = "^2.2.0"
httpx = "^0.27.0"

# Auth & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"  # Pin to 4.0.x for passlib compatibility

# Utilities
pydantic = "^2.8.0"
pydantic-settings = "^2.3.0"
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"

# PDF Generation
reportlab = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
ruff = "^0.5.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
